requirements.txt
catboost==1.2.5
easyocr==1.7.1
fastapi==0.111.0
gdown==5.2.0
numpy==1.26.4
onnxruntime==1.17.3
onnx==1.16.0
Pillow==10.3.0
scipy==1.13.0
torch==2.3.0
torchvision==0.18.0
transformers==4.40.2
pytest==8.2.0